[feature] update home

+++ app/src/main/java/kim/bifrost/rain/rainmusic/model/web/bean/netease/NeteaseLoginBean.kt
@@ -60,7 +60,7 @@ data class NeteaseLoginBean(
         val backgroundImgId: Long,
         val backgroundImgIdStr: String,
         val backgroundUrl: String,
-        val birthday: Int,
+        val birthday: Long,
         val city: Int,
         val defaultAvatar: Boolean,
         val description: String,
